requirements.txt
# CIVI-GENESIS Dependencies
# Python 3.10+

# Core framework
streamlit>=1.28.0

# Data manipulation
pandas>=2.0.0
numpy>=1.24.0

# Machine learning
scikit-learn>=1.3.0
joblib>=1.3.0

# Visualization
plotly>=5.17.0
altair>=5.0.0

# Google Gemini API
google-generativeai>=0.3.0

# Utilities
python-dotenv>=1.0.0
